requirements.txt
aiogram==2.25.1
ccxt==4.1.0
aioschedule==0.5.2
python-dotenv==1.0.0